Move utility functions to new module
